Store event in settings context to persist it
